Move memex.presenters into h.presenters
Move the various Annotation and Document presenter classes into the
newly-created `h.presenters` package. No presenter code has changed
here, so the apparent size of the diff is a little misleading. The
presenters have been split up into individual modules so that there is
one presenter class defined by each module in `h.presenters`.

To assist making this move the following things have had to change at
the same time:

- the 'json' and 'jsonld_id' link definitions have moved into `h.links`
- the 'utc_iso8601' helper function has moved into `h.util.datetime`
- Jinja2 has been added to the list of memex dependencies. This is
  because I've had to introduce a reverse dependency from memex on the
  `h.presenters` module, which now includes HTML presenters, and so we
  need Jinja2 in the memex test environment.
