+++ h/presenters/__init__.py
@@ -8,9 +8,19 @@ formats.
 from __future__ import unicode_literals
 
 from h.presenters.annotation_html import AnnotationHTMLPresenter
+from h.presenters.annotation_json import AnnotationJSONPresenter
+from h.presenters.annotation_jsonld import AnnotationJSONLDPresenter
+from h.presenters.annotation_searchindex import AnnotationSearchIndexPresenter
 from h.presenters.document_html import DocumentHTMLPresenter
+from h.presenters.document_json import DocumentJSONPresenter
+from h.presenters.document_searchindex import DocumentSearchIndexPresenter
 
 __all__ = (
     'AnnotationHTMLPresenter',
+    'AnnotationJSONPresenter',
+    'AnnotationJSONLDPresenter',
+    'AnnotationSearchIndexPresenter',
     'DocumentHTMLPresenter',
+    'DocumentJSONPresenter',
+    'DocumentSearchIndexPresenter',
 )
