+++ tests/h/views/api_test.py
@@ -6,11 +6,11 @@ import pytest
 from pyramid.config import Configurator
 from pyramid import testing
 
-from memex import presenters
 from memex.resources import AnnotationResource
 from memex.schemas import ValidationError
 from memex.search.core import SearchResult
 
+from h import presenters
 from h.views import api as views
